backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.1.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.4.0
google-genai==1.1.0
